+++ Docs/releases.md
@@ -1,7 +1,10 @@
 # Releases
 
+## Version 1.2
+- Fix KeyNotFoundException when resolving closed generic types through intermediate abstract classes ([#6](https://github.com/mrpmorris/AutoRegister/issues/6))
+
 ## Version 1.1
-- Allow service keys to be declared in a different assembly ([#4](https://github.com/mrpmorris/AutoRegister/issues/4))
+- Allow service keys (types) to be declared in a different assembly ([#4](https://github.com/mrpmorris/AutoRegister/issues/4))
 
 ## Version 1.0
 First production-ready release.